Moved FileDropHandle to a separate file.
